+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Tamed pets now get the beacon effects
+
 ## 2.12.7
 * Well Fed and Injured can now be configured (Duration and effectiveness)
   * Halved Injured duration